Move duplicated code to a base class for group_mapping migrations
