+++ apps/user_ldap/lib/Migration/Version1130Date20220110154718.php
@@ -28,48 +28,12 @@ namespace OCA\User_LDAP\Migration;
 
 use Closure;
 use OCP\DB\ISchemaWrapper;
+use OCP\DB\Types;
 use OCP\Migration\IOutput;
-use OCP\Migration\SimpleMigrationStep;
 
-class Version1130Date20220110154718 extends SimpleMigrationStep {
-	/** @var IDBConnection */
-	private $dbc;
-
-	public function __construct(IDBConnection $dbc) {
-		$this->dbc = $dbc;
-	}
-
+class Version1130Date20220110154718 extends GroupMappingMigration {
 	public function getName() {
 		return 'Copy ldap_group_mapping data from backup table and if needed';
-	}
-
-	protected function copyGroupMappingData(string $sourceTable, string $destinationTable): void {
-		$insert = $this->dbc->getQueryBuilder();
-		$insert->insert($destinationTable)
-			->values([
-				'ldap_dn' => $insert->createParameter('ldap_dn'),
-				'owncloud_name' => $insert->createParameter('owncloud_name'),
-				'directory_uuid' => $insert->createParameter('directory_uuid'),
-				'ldap_dn_hash' => $insert->createParameter('ldap_dn_hash'),
-			]);
-
-		$query = $this->dbc->getQueryBuilder();
-		$query->select('*')
-			->from($sourceTable);
-
-
-		$result = $query->executeQuery();
-		while ($row = $result->fetch()) {
-			$insert
-				->setParameter('ldap_dn', $row['ldap_dn'])
-				->setParameter('owncloud_name', $row['owncloud_name'])
-				->setParameter('directory_uuid', $row['directory_uuid'])
-				->setParameter('ldap_dn_hash', $row['ldap_dn_hash'])
-				;
-
-			$insert->executeStatement();
-		}
-		$result->closeCursor();
 	}
 
 	/**
